folder structure fix

+++ LoanApplication/Services/Loan/ILoanService.cs
@@ -7,6 +7,6 @@ public interface ILoanService
 {
     Task<Loan?> GetLoanByIdAsync(int loanId);
     Task<IEnumerable<Loan>> GetAllLoansAsync();
-    Task<IEnumerable<Loan>> GetAllLoansByUserIdAsync(long userId);
+    Task<IEnumerable<Loan>> GetAllLoansByUserIdAsync(int userId);
     Task<Loan> CreateLoanAsync(Loan loan);
 }
